feat: 변수 타입 변경 및 엔티티 생성 메소드 추가

+++ src/main/java/com/dbd/seoulcinema/domain/entity/Ticket.java
@@ -30,7 +30,7 @@ public class Ticket extends BaseTimeEntity {
     private TicketingStatus ticketingStatus;
 
     @Column(name = "STANDARD_PRICE")
-    private Long standardPrice;
+    private Integer standardPrice;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "CLIENT_NUMBER")
